Start API stub with Gradle
