+++ .gitignore
@@ -4,6 +4,7 @@
 !.yarn/plugins
 !.yarn/sdks
 !.yarn/versions
+apps/challenge-user-service-2
 
 # See http://help.github.com/ignore-files/ for more about ignoring files.
 .coveralls.yml
