--- apps/challenge-user-service/server/src/main/java/org/sagebionetworks/challenge/model/dto/UserUpdateRequest.java
@@ -1,123 +1,0 @@
-package org.sagebionetworks.challenge.model.dto;
-
-import java.net.URI;
-import java.util.Objects;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-import org.openapitools.jackson.nullable.JsonNullable;
-import java.time.OffsetDateTime;
-import javax.validation.Valid;
-import javax.validation.constraints.*;
-import io.swagger.v3.oas.annotations.media.Schema;
-
-
-import java.util.*;
-import javax.annotation.Generated;
-
-/**
- * UserUpdateRequest
- */
-
-@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
-public class UserUpdateRequest {
-
-  /**
-   * Gets or Sets status
-   */
-  public enum StatusEnum {
-    PENDING("PENDING"),
-    
-    APPROVED("APPROVED"),
-    
-    DISABLED("DISABLED"),
-    
-    BLACKLIST("BLACKLIST");
-
-    private String value;
-
-    StatusEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static StatusEnum fromValue(String value) {
-      for (StatusEnum b : StatusEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
-  @JsonProperty("status")
-  private StatusEnum status;
-
-  public UserUpdateRequest status(StatusEnum status) {
-    this.status = status;
-    return this;
-  }
-
-  /**
-   * Get status
-   * @return status
-  */
-  
-  @Schema(name = "status", required = false)
-  public StatusEnum getStatus() {
-    return status;
-  }
-
-  public void setStatus(StatusEnum status) {
-    this.status = status;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    UserUpdateRequest userUpdateRequest = (UserUpdateRequest) o;
-    return Objects.equals(this.status, userUpdateRequest.status);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(status);
-  }
-
-  @Override
-  public String toString() {
-    StringBuilder sb = new StringBuilder();
-    sb.append("class UserUpdateRequest {\n");
-    sb.append("    status: ").append(toIndentedString(status)).append("\n");
-    sb.append("}");
-    return sb.toString();
-  }
-
-  /**
-   * Convert the given object to string with each line indented by 4 spaces
-   * (except the first line).
-   */
-  private String toIndentedString(Object o) {
-    if (o == null) {
-      return "null";
-    }
-    return o.toString().replace("\n", "\n    ");
-  }
-}
-
