+++ apps/challenge-user-service/src.old/integrationTest/java/org/sagebionetworks/challenge/ChallengeUserServiceApplicationTests.java
