+++ apps/challenge-user-service/src.old/integrationTest/java/org/sagebionetworks/challenge/model/repository/UserRepositoryIntegrationTest.java
