+++ apps/challenge-user-service/src.old/integrationTest/resources/application.yml
