+++ apps/challenge-user-service/src.old/main/java/org/sagebionetworks/challenge/configuration/CustomFeignClientConfiguration.java
