+++ apps/challenge-user-service/src.old/main/java/org/sagebionetworks/challenge/model/dto/UserUpdateRequest.java
@@ -1,0 +1,8 @@
+package org.sagebionetworks.challenge.model.dto;
+
+import lombok.Data;
+
+@Data
+public class UserUpdateRequest {
+  private UserStatus status;
+}
