+++ apps/challenge-user-service/src.old/main/java/org/sagebionetworks/challenge/model/repository/UserRepository.java
