+++ apps/challenge-user-service/src.old/main/java/org/sagebionetworks/challenge/model/rest/response/UserResponse.java
