+++ apps/challenge-user-service/src.old/main/resources/db/migration/V1.0.20210427174638__create_user_table.sql
