+++ apps/challenge-user-service/src.old/main/resources/db/migration/V1.0.20210427174721__temp_data.sql
