+++ apps/challenge-user-service/src.old/test/java/org/sagebionetworks/challenge/controller/UserControllerTest.java
