+++ apps/challenge-user-service/src.old/test/java/org/sagebionetworks/challenge/service/UserServiceTest.java
