+++ apps/challenge-user-service/src.old/test/resources/application.yml
