+++ apps/challenge-user-service/src/main/java/org/sagebionetworks/challenge/api/UserControllerApiDelegate.java
