+++ apps/challenge-user-service/src/main/java/org/sagebionetworks/challenge/configuration/HomeController.java
