+++ apps/challenge-user-service/src/main/resources/application.properties
