+++ apps/challenge-user-service/src/main/resources/openapi.yaml
