+++ apps/challenge-user-service/src/test/java/org/sagebionetworks/challenge/OpenApiGeneratorApplicationTests.java
